% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{anova.mcglm}
\alias{anova.mcglm}
\title{ANOVA method for McGLMs.}
\usage{
\method{anova}{mcglm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{mcglm}, usually, a result of a
call to \code{mcglm()}.}

\item{...}{additional arguments affecting the summary produced. Note
that there is no extra options for mcglm object class.}
}
\value{
A \code{data.frame} with Chi-square statistic to test the
    null hypothesis of a parameter, or a set of parameters, be
    zero. The Wald test based on the observed covariance matrix of
    the parameters is used.
}
\description{
ANOVA method for object of class McGLMS.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

