% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_confint.mcglm.R
\name{confint.mcglm}
\alias{confint.mcglm}
\title{Confidence Intervals for mcglm}
\usage{
\method{confint}{mcglm}(object, level = 0.95)
}
\arguments{
\item{object}{a fitted mcglm object.}

\item{level}{the confidence level required.}
}
\value{
<<<<<<< HEAD
A data.frame with confidence intervals, parameters names,
response number and parameters type.
=======
A data.frame with confidence intervals, parameters names, response number and parameters type.
>>>>>>> b258eea0594d0aef2a35c6f727b6b001959cfae1
}
\description{
Computes confidence intervals for parameters in a fitted mcglm.
}

