% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_ess.R
\name{ESS}
\alias{ESS}
\title{Extract generalized error sum of squares (ESS) for multivariate
    covariance generalized linear models.}
\usage{
ESS(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{verbose}{Logical}
}
\value{
Returns the value of the Gaussian pseudo-loglikelihood.
}
\description{
Extract the generalized error sum of squares for a fitted
McGLM.
}
\author{
Wagner Hugo Bonat
}

