% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_fitted.mcglm.R
\name{fitted.mcglm}
\alias{fitted.mcglm}
\title{Extract Model Fitted Values of McGLM}
\usage{
\method{fitted}{mcglm}(object)
}
\arguments{
\item{object}{An object of mcglm class.}
}
\value{
Depending on the number of response variable the function \code{fitted.mcglm} returns
a vector (univariate models) or a matrix (multivariate models) of fitted values.
}
\description{
Extract fitted values for objects of mcglm class.
}

