% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_gof.R
\name{gof}
\alias{gof}
\title{Extract measures of goodness-of-fit for
multivariate covariance generalized linear models.}
\usage{
gof(object)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}
}
\value{
Returns a data frame contains gof's.
}
\description{
Extract the plogLik, pAIC, pKLIC and ESS for a fitted McGLM.
}
\author{
Wagner Hugo Bonat
}

