% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_build_bdiag.R
\name{mc_build_bdiag}
\alias{mc_build_bdiag}
\title{Build a block-diagonal matrix of zeros.}
\usage{
mc_build_bdiag(n_resp, n_obs)
}
\arguments{
\item{n_resp}{A numeric specifyng the number of response variables.}

\item{n_obs}{A numeric specifying the number of observations in the
data set.}
}
\value{
A list of zero matrices.
}
\description{
Build a block-diagonal matrix of zeros. Such functions
    is used when computing the derivatives of the Cholesky
    decomposition of C.
}
\details{
It is an internal function.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}

