% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_build_F.R
\name{mc_build_F}
\alias{mc_build_F}
\title{Auxiliar function: Build F matrix for Wald multivariate test}
\usage{
mc_build_F(vector)
}
\arguments{
\item{vector}{A vector indexing model regression coefficients.}
}
\value{
A matrix.
}
\description{
The function \code{mc_build_F} is just an auxiliar
    function to construct the design matrix for multivariate Wald-type
    test.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
