% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_build_omega.R
\name{mc_build_omega}
\alias{mc_build_omega}
\title{Build omega matrix}
\usage{
mc_build_omega(tau, Z, covariance_link, sparse = FALSE)
}
\arguments{
\item{tau}{A vector}

\item{Z}{A list of matrices.}

\item{covariance_link}{String specifing the covariance link function:
identity, inverse, expm.}

\item{sparse}{Logical force to use sparse matrix representation
'dsCMatrix'.}
}
\value{
A list with the \eqn{\Omega} matrix its inverse and
    derivatives with respect to \eqn{\tau}.
}
\description{
This function build \eqn{\Omega} matrix according the
    covariance link function.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
