% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_car.R
\name{mc_car}
\alias{mc_car}
\title{Conditional autoregressive models}
\usage{
mc_car(list_neigh, intrinsic = FALSE)
}
\arguments{
\item{list_neigh}{List of neighboors.}

\item{intrinsic}{Logical}
}
\value{
A matrix.
}
\description{
Builds conditional autoregressive model matrix.
}

