% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_compute_rho.R
\name{mc_compute_rho}
\alias{mc_compute_rho}
\title{Compute autocorrelation estimates}
\usage{
mc_compute_rho(object, level = 0.975)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{level}{the confidence level required.}
}
\value{
Returns estimate, standard error and confidential interval for
the spatial autocorrelation parameter.
}
\description{
Compute autocorrelation estimates based on models fitting
using the \code{mc_car} structure.
}
\author{
Wagner Hugo Bonat
}

