% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_conditional_test.R
\name{mc_conditional_test}
\alias{mc_conditional_test}
\title{Compute conditional hypotheses test for multivariate covariance
generalized linear models}
\usage{
mc_conditional_test(object, parameters, test, fixed)
}
\arguments{
\item{object}{an object representing a model of \code{mcglm} class.}

\item{parameters}{Which parameters have been included in the
conditional test.}

\item{test}{Index indicating which parameters will be tested given
the values of the other parameters.}

\item{fixed}{Index indicating which parameters should be fixed on
the conditional test.}
}
\value{
Returns estimates, conditional standard errors and Z statistics.
}
\description{
Compute conditional hypotheses test for fitted McGLMs.
It is useful in general when estimating the power parameter.
}
\author{
Wagner Hugo Bonat
}

