% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_core_pearson.R
\name{mc_core_pearson}
\alias{mc_core_pearson}
\title{Core of the Pearson estimating function.}
\usage{
mc_core_pearson(product, inv_C, res)
}
\arguments{
\item{product}{A matrix.}

\item{inv_C}{A matrix.}

\item{res}{A vector of residuals.}
}
\value{
A vector
}
\description{
Core of the Pearson estimating function.
}
\details{
It is an internal function.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
