% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_cross_variability.R
\name{mc_cross_variability}
\alias{mc_cross_variability}
\title{Compute the cross-variability matrix}
\usage{
mc_cross_variability(Product_cov, inv_C, res, D)
}
\arguments{
\item{Product_cov}{A list of matrices.}

\item{inv_C}{A matrix.}

\item{res}{A vector.}

\item{D}{A matrix.}
}
\value{
The cross-variability matrix between regression and
    covariance parameters.
}
\description{
Compute the cross-variability matrix between covariance
    and regression parameters.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}

