% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_derivative_C_rho.R
\name{mc_derivative_C_rho}
\alias{mc_derivative_C_rho}
\title{Derivative of C with respect to rho.}
\usage{
mc_derivative_C_rho(
  D_Sigmab,
  Bdiag_chol_Sigma_within,
  t_Bdiag_chol_Sigma_within,
  II
)
}
\arguments{
\item{D_Sigmab}{A matrix.}

\item{Bdiag_chol_Sigma_within}{A block-diagonal matrix.}

\item{t_Bdiag_chol_Sigma_within}{A block-diagonal matrix.}

\item{II}{A diagonal matrix.}
}
\value{
A matrix.
}
\description{
Compute the derivative of the C matrix with respect to
    the correlation parameters rho.
}
\details{
It is an internal function used to build the derivatives of
    the C matrix.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
