% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_derivative_cholesky.R
\name{mc_derivative_cholesky}
\alias{mc_derivative_cholesky}
\title{Derivatives of the Cholesky decomposition}
\usage{
mc_derivative_cholesky(derivada, inv_chol_Sigma, chol_Sigma)
}
\arguments{
\item{derivada}{A matrix.}

\item{inv_chol_Sigma}{A matrix.}

\item{chol_Sigma}{A matrix.}
}
\value{
A list of matrix. The matrices are the derivative of the Cholesky decomposition
of Sigma or its inverse with respect to the dispersion parameters.
The returned object is intended for internal use only.
}
\description{
This function compute the derivative of the Cholesky
    decomposition.
}
\details{
It is an internal function.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
