% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_influence.R
\name{mc_dfbetaOij}
\alias{mc_dfbetaOij}
\title{Influence measures for McGLMs}
\usage{
mc_dfbetaOij(Di, Ci, inv_Ci, ri, inv_M)
}
\arguments{
\item{Di}{D matrix for the cluster i.}

\item{Ci}{C matrix for the cluster i.}

\item{inv_Ci}{Inverse of C matrix for the cluster i.}

\item{ri}{Residual vector for the cluster i.}

\item{inv_M}{Inverse of variance/covariance of regression parameters.}
}
\value{
Matrix with the DFBETA for observation in the cluster i.
}
\description{
Compute influence measures for multivariate covariance
generalized linear models. Leverage, DFBETA and Cook's distance
for observations. Auxiliar function for \code{mc_influence}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

