% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_dist.R
\name{mc_dist}
\alias{mc_dist}
\title{Distances models}
\usage{
mc_dist(id, time, data, method = "euclidean")
}
\arguments{
\item{id}{Subject index. Note that this structure was designed to deal with longitudinal data.}

\item{time}{Index indicating the time.}

\item{data}{Data set}

\item{method}{The distance measure to be used. This must be one of
"euclidean", "maximum","manhattan","canberra","binary" or "minkowski".}
}
\value{
A matrix of dgCMatrix class.
}
\description{
Builds distances model matrix.
}

