% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_fast_forward.R
\name{mc_fast_forward}
\alias{mc_fast_forward}
\title{Fast forward selection for multivariate covariance generalized linear
models.}
\usage{
mc_fast_forward(object, scope, interaction = 1, penalty = 2, n_max = 10)
}
\arguments{
\item{object}{an object representing a model of \code{mcglm} class.}

\item{scope}{a vector specyfing the covariate to be tested.}

\item{interaction}{Maximum number of covariates interacting.}

\item{penalty}{penalty term (default = 2).}
}
\value{
The selected model.
}
\description{
Perform fast forward model selection using the score
information criterion. This function works only for univariate months.
}

