% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_id.R
\name{mc_id}
\alias{mc_id}
\title{Independent model structure}
\usage{
mc_id(data)
}
\arguments{
\item{data}{The data set to be used.}
}
\value{
A list of matrix.
}
\description{
Builds a identity matrix to be used as a component of
the matrix linear predictor.
}

