% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_influence.R
\name{mc_influence}
\alias{mc_influence}
\title{Influence measures for McGLMs}
\usage{
mc_influence(object, id)
}
\arguments{
\item{object}{An object of \code{mcglm} class.}

\item{id}{a vector which identifies the clusters.
The length and order of id should match with the number of
observations.}
}
\value{
A list with influence measures for cluster and observations.
}
\description{
Compute influence measures for multivariate covariance
generalized linear models. Leverage, DFBETA and Cook's distance
for unit sample and observations.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

