% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_list2vec.R
\name{mc_list2vec}
\alias{mc_list2vec}
\title{Auxiliar function transforms list to a vector.}
\usage{
mc_list2vec(list_initial, list_power_fixed)
}
\arguments{
\item{list_initial}{A list specifying initial values.}

\item{list_power_fixed}{A list of logical operators specyfing if the
power parameter should be estimated or not.}
}
\value{
A list with the following components:
\describe{
  \item{beta_ini}{A numeric vector containing the initial values of the
  regression parameters stacked across all response variables.}
  \item{cov_ini}{A numeric vector containing the initial values of the
  covariance-related parameters, including dispersion, power, and
  correlation parameters, after removing fixed power parameters.}
}
}
\description{
This function takes a list of parameters and tranforms
    to a vector.
}
\details{
It is an internal function, in general the users never will
    use this function.  It will be useful, only if the user wants to
    implement a different variance-covariance matrix.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
