% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_ma.R
\name{mc_ma}
\alias{mc_ma}
\title{Moving average models}
\usage{
mc_ma(id, time, data, order = 1)
}
\arguments{
\item{id}{Subject index. Note that this structure was designed to deal with longitudinal data.}

\item{time}{Index indicating the time.}

\item{data}{Data set.}

\item{order}{Order of the random walk model.}
}
\value{
A matrix of dgCMatrix class.
}
\description{
Builds moving average model matrix of order k.
}

