% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_matrix_linear_predictor.R
\name{mc_matrix_linear_predictor}
\alias{mc_matrix_linear_predictor}
\title{Matrix linear predictor}
\usage{
mc_matrix_linear_predictor(tau, Z)
}
\arguments{
\item{tau}{A numeric vector.}

\item{Z}{A list of known matrices.}
}
\value{
A matrix.
}
\description{
Compute the matrix linear predictor.
}
\details{
Given a list with a set of known matrices
    (\eqn{Z_0,...,Z_D}) the function
    \code{mc_matrix_linear_predictor} returns \eqn{U = \tau_0 Z_0 +
    ... + \tau_D Z_D}.
}
\examples{
require(Matrix)
Z0 <- Diagonal(5, 1)
Z1 <- Matrix(rep(1,5)\%*\%t(rep(1,5)))
Z <- list(Z0, Z1)
mc_matrix_linear_predictor(tau = c(1,0.8), Z = Z)
}
\author{
Wagner Hugo Bonat
}

