% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_mixed.R
\name{mc_mixed}
\alias{mc_mixed}
\title{Mixed model structure}
\usage{
mc_mixed(formula, data)
}
\arguments{
\item{formula}{A formula model to build the matrix linear predictor (see Details).}

\item{data}{The data set to be used.}
}
\value{
A list matrices.
}
\description{
Builds a mixed model structure
}

