% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_pearson.R
\name{mc_pearson}
\alias{mc_pearson}
\title{Pearson estimating function}
\usage{
mc_pearson(y_vec, mu_vec, Cfeatures, inv_J_beta = NULL, D = NULL,
  correct = FALSE, compute_variability = FALSE)
}
\arguments{
\item{y_vec}{A vector.}

\item{mu_vec}{A vector.}

\item{inv_J_beta}{A matrix.}

\item{D}{A matrix.}

\item{correct}{Logical.}

\item{compute_variability}{Logical.}

\item{A}{list of matrices.}
}
\description{
Compute the Pearson estimating function its sensitivity and variability matrices.
}
\details{
Compute the Pearson estimating function its sensitivity and variability matrices.
For more details see Bonat and Jorgensen (2015) equations 6-8.
}

