% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_qll.R
\name{mc_qll}
\alias{mc_qll}
\title{Compute quasi-likelihood function.}
\usage{
mc_qll(y, mu, variance, power)
}
\arguments{
\item{y}{A vector of observed values.}

\item{mu}{A vector of fitted values.}

\item{variance}{Variance function (constant, tweedie, poisson_tweedie, binomial).}

\item{power}{Power parameter value.}
}
\value{
The quasi-likelihood values.
}
\description{
Given a variance function mc_qll function computes the quasi-likelihood values.
}

