% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_robust_std.R
\name{mc_robust_std}
\alias{mc_robust_std}
\title{Robust standard error for regression parameters}
\usage{
mc_robust_std(object, id)
}
\arguments{
\item{object}{An object of mcglm class.}

\item{id}{a vector which identifies the clusters. The length and order of id should be the
same as the number of observations. Data are assumed to be sorted so that observations on a cluster
are contiguous rows for all entities in the formula.}
}
\value{
A matrix. Note that the function assumes that the data are in the correct order.
}
\description{
Compute robust standard error for regression parameters in the context of clustered
observations.
}

