% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rw.R
\name{mc_rw}
\alias{mc_rw}
\title{Random walk models}
\usage{
mc_rw(id, time, data, order = 1, proper = FALSE)
}
\arguments{
\item{id}{Subject index. Note that this structure was designed to deal with longitudinal data.}

\item{time}{Index indicating the time.}

\item{data}{Data set.}

\item{order}{Order of the random walk model.}

\item{proper}{Logical.}
}
\value{
A matrix of dgCMatrix class.
}
\description{
Builds a random walk model matrix of order k.
}

