% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_rw1.R
\name{mc_rw1}
\alias{mc_rw1}
\title{Random walk first order model}
\usage{
mc_rw1(n_time, intrinsic = TRUE)
}
\arguments{
\item{n_time}{Number observations time.}

\item{intrinsic}{Logical indicating if the models is intrinsic (rho = 1) or not.}
}
\value{
A matrix. Note that the function assumes that the data are in the correct order.
}
\description{
Builds a random walk first order model matrix.
}

