% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_rw2.R
\name{mc_rw2}
\alias{mc_rw2}
\title{Random walk second order model}
\usage{
mc_rw2(n_time, intrinsic = TRUE)
}
\arguments{
\item{n_time}{Number observations time.}

\item{intrinsic}{Logical indicating if the models is intrinsic (rho = 1) or not.}
}
\value{
A matrix. Note that the function assumes that the data are in the correct order.
}
\description{
Builds a random walk second order model matrix.
}

