% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sic.R
\name{mc_sic}
\alias{mc_sic}
\title{Compute the score information criterion (SIC) for multivariate
    covariance generalized linear models.}
\usage{
mc_sic(object, scope, data, response, penalty = 2)
}
\arguments{
\item{object}{an object representing a model of \code{mcglm} class.}

\item{scope}{a vector containing all covariate names to be tested.}

\item{data}{data frame containing the all variables envolved}

\item{response}{Indicate for which response variable SIC is computed.}

\item{penalty}{penalty term (default = 2).}
}
\value{
A data frame with SIC values for each covariate in the scope
    argument.
}
\description{
Compute the SIC for McGLMS.
}
\author{
Wagner Hugo Bonat
}

