% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_sic_covariance.R
\name{mc_sic_covariance}
\alias{mc_sic_covariance}
\title{Compute the score information criterion (SIC) for multivariate
    covariance generalized linear models.}
\usage{
mc_sic_covariance(object, scope, idx, data, penalty = 2, response)
}
\arguments{
\item{object}{an object representing a model of \code{mcglm} class.}

\item{scope}{a list of matrices to be tested in the matrix linear
predictor.}

\item{idx}{Indicator of matrices belong to the same effect.}

\item{data}{data frame containing all variables envolved in the
model.}

\item{penalty}{penalty term (default = 2).}

\item{response}{Indicate for which response variable SIC is computed.}
}
\value{
A data frame with SIC values for each matrix in the scope
    argument.
}
\description{
Compute SIC for covariance parameters in McGLMS.
}
\author{
Wagner Hugo Bonat
}

