% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_unstructured.R
\name{mc_unstructured}
\alias{mc_unstructured}
\title{Unstructured model}
\usage{
mc_unstructured(n_time)
}
\arguments{
\item{n_time}{Number of observations per unit sample.}
}
\value{
A matrix. Note that the function assumes that the data are in the correct order.
}
\description{
Builds a unstructured model matrix.
}

