% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_main_function.R, R/mcglm-package.r, R/mcglm.R
\docType{package}
\name{mcglm}
\alias{mcglm}
\alias{mcglm-package}
\title{Fitting Multivariate Covariance Generalized Linear Models.}
\source{
Bonat, W.H. and Jorgensen, B. (2016). Multivariate
    covariance generalized linear models. Journal of Royal Statistical
    Society - Series C. to appear.
}
\usage{
mcglm(linear_pred, matrix_pred, link, variance, covariance, offset, Ntrial,
  power_fixed, data, control_initial = "automatic", contrasts = NULL,
  control_algorithm = list())
}
\arguments{
\item{linear_pred}{A list of formula see \code{\link[stats]{formula}}
for details.}

\item{matrix_pred}{A list of known matrices to be used on the matrix
linear predictor. Details can be obtained on
\code{\link[mcglm]{mc_matrix_linear_predictor}}.}

\item{link}{A list of link functions names, see
\code{\link[mcglm]{mc_link_function}} for details.}

\item{variance}{A list of variance functions names, see
\code{\link[mcglm]{mc_variance_function}} for details.}

\item{covariance}{A list of covariance link functions names, current
options are: identity, inverse and exponential-matrix (expm).}

\item{offset}{A list with values of offset values if any.}

\item{Ntrial}{A list with values of the number of trials on Bernoulli
experiments. It is useful only for binomialP and binomialPQ
variance functions.}

\item{power_fixed}{A list of logicals indicating if the values of the
power parameter should be estimated or not.}

\item{data}{A dta frame.}

\item{control_initial}{A list of initial values for the fitting
algorithm. See details below.}

\item{contrasts}{Extra arguments to passed to
\code{\link[stats]{model.matrix}}.}

\item{control_algorithm}{A list of arguments to be passed for the
fitting algorithm. See \code{\link[mcglm]{fit_mcglm}} for
details.}
}
\value{
mcglm returns an object of class 'mcglm'.
}
\description{
The function \code{mcglm} is used to fit multivariate
    covariance generalized linear models.
    The models are specified by a set of lists giving a symbolic
    description of the linear predictor.
    The user can choose between a list of link, variance and covariance
    functions. The models are fitted using an estimating function
    approach, combining quasi-score functions for regression
    parameters and Pearson estimating function for covariance
    parameters. For details see Bonat and Jorgensen (2016).

mcglm.

This package fits multivariate covariance generalized
    linear models (McGLMs) to data.  McGLM is a general framework for
    non-normal multivariate data analysis, designed to handle
    multivariate response variables, along with a wide range of
    temporal and spatial correlation structures defined in terms of
    a covariance link function combined with a matrix linear
    predictor involving known matrices. The models take
    non-normality into account in the conventional way by means of a
    variance function, and the mean structure is modelled by means
    of a link function and a linear predictor.  The models are
    fitted using an efficient Newton scoring algorithm based on
    quasi-likelihood and Pearson estimating functions, using only
    second-moment assumptions.  This provides a unified approach to
    a wide variety of different types of response variables and
    covariance structures, including multivariate extensions of
    repeated measures, time series, longitudinal, spatial and
    spatio-temporal structures. The package offers a user-friendly
    interface for fitting McGLMS.
}
\author{
Wagner Hugo Bonat
}

