% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_pAIC.R
\name{pAIC}
\alias{pAIC}
\title{Extract pseudo Akaike Information Criterion (pAIC) for
multivariate covariance generalized linear models.}
\usage{
pAIC(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{verbose}{Logical}
}
\value{
Returns the value of the pAIC.
}
\description{
Extract the pAIC for a fitted McGLM.
}
\author{
Wagner Hugo Bonat
}

