% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_pBIC.R
\name{pBIC}
\alias{pBIC}
\title{Pseudo Bayesian Information Criterion}
\source{
Bonat, W. H. (2018). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, 84(4):1--30.
}
\usage{
pBIC(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{verbose}{logical. Print or not the pBIC value.}
}
\value{
Returns the value of the pseudo Bayesian information criterion (pBIC).
}
\description{
Extract the pseudo Bayesian information criterion (pBIC)
for objects of \code{mcglm} class.
}
\seealso{
\code{gof}, \code{plogLik}, \code{ESS}, \code{pKLIC},
\code{GOSHO} and \code{RJC}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
