% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_KLIC.R
\name{pKLIC}
\alias{pKLIC}
\title{Extract pseudo Kullback-Leibler Information Criterion for
multivariate covariance generalized linear models.}
\usage{
pKLIC(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{verbose}{Logical}
}
\value{
Returns the value of the pKLIC.
}
\description{
Extract the pKLIC for a fitted McGLM.
}
\author{
Wagner Hugo Bonat
}

