% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_plogLik.R
\name{plogLik}
\alias{plogLik}
\title{Extract Gaussian pseudo-loglikelihood (plogLik) for multivariate
    covariance generalized linear models.}
\usage{
plogLik(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{verbose}{Logical}
}
\value{
Returns the value of the Gaussian pseudo-loglikelihood.
}
\description{
Extract the Gaussian pseudo-loglikelihood for a fitted
McGLM.
}
\author{
Wagner Hugo Bonat
}

