% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{plot.mcglm}
\alias{plot.mcglm}
\title{Diagnostic Plots for mcglm Objects}
\usage{
\method{plot}{mcglm}(x, type = c("residuals", "algorithm", "partial_residuals"), ...)
}
\arguments{
\item{x}{A fitted object of class \code{mcglm}.}

\item{type}{Character string specifying the type of diagnostic plot to be produced.
Possible values are \code{"residuals"}, \code{"algorithm"} and
\code{"partial_residuals"}.}

\item{...}{Additional arguments. Currently ignored and included for compatibility
with the generic \code{\link[graphics]{plot}} function.}
}
\value{
No return value, called for its side effects (diagnostic plots).
}
\description{
Produces diagnostic plots for fitted \code{mcglm} objects.
Available diagnostics include residual analysis, inspection of
the fitting algorithm convergence and partial residual plots.
}
\seealso{
\code{\link[stats]{residuals}},
\code{\link[stats]{fitted}},
\code{\link[graphics]{plot}}
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
