% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_print.mcglm.R
\name{print.mcglm}
<<<<<<< HEAD
\alias{print}
=======
>>>>>>> b258eea0594d0aef2a35c6f727b6b001959cfae1
\alias{print.mcglm}
\title{Print a Multivariate Covariance Generalized Linear Model}
\usage{
print.mcglm(object)
}
\arguments{
\item{object}{fitted model objects of class mcglm as produced by mcglm().}
}
\description{
The default print method for a mcglm object.
}

