% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_qic.R
\name{qic.mcglm}
\alias{qic.mcglm}
\title{Compute Quasi Information Criterion (QIC) for McGLMs.}
\usage{
qic.mcglm(object, object.iid)
}
\arguments{
\item{object}{An object of mcglm class.}

\item{object.iid}{An object of mcglm class contained the model fitted using independent
covariance structure.}
}
\value{
The QIC value.
}
\description{
qic.mcglm is a function which computes the QIC for McGLMs.
}

