% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_RJC.R
\name{RJC}
\alias{RJC}
\title{Rotnitzky-Jewell information criterion}
\usage{
RJC(object, id, verbose = TRUE)
}
\arguments{
\item{object}{An object of mcglm class.}

\item{id}{a vector which identifies the clusters. The length and
order of id should be the same as the number of
observations. Data are assumed to be sorted so that observations
on a cluster are contiguous rows for all entities in the formula.}

\item{verbose}{Logical.}
}
\value{
A matrix. Note that the function assumes that the data are in
    the correct order.
}
\description{
Compute the Rotnitzky-Jewell information criterion
for multivariate covariance generalized linear models.
WARNINGS: This function is limited to models with ONE response variable.
}
\author{
Wagner Hugo Bonat
}

