% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_updatedCov.R
\name{updatedCov}
\alias{updatedCov}
\title{Updated covariance parameters}
\usage{
updatedCov(list_initial, covariance, list_power_fixed, information, n_resp)
}
\arguments{
\item{list_initial}{A list of initial values.}

\item{covariance}{A vector with actual covariance parameters values.}

\item{list_power_fixed}{A list of logical indicating if the power parameter should be estimated or not.}

\item{information}{A list with information about the number of parameters in the model. In general
the output from \code{[mcglm]{getInformation}}.}

\item{n_resp}{A numeric specyfing the number of response variables.}
}
\value{
A list with updated values of the covariance parameters.
}
\description{
This function update a list of covariance parameters. It will be useful only inside the
fitting algorithm.
}

