% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{vcov.mcglm}
\alias{vcov.mcglm}
\title{Calculate Variance-Covariance matrix for a fitted McGLM
    object.}
\usage{
\method{vcov}{mcglm}(object, ...)
}
\arguments{
\item{object}{a fitted model \code{mcglm} object.}

\item{...}{additional arguments affecting the summary produced. Note
    that there is no extra options for mcglm object class.}
}
\value{
A variance-covariance matrix.
}
\description{
Returns the variance-covariance matrix for all
    parameters of a \code{mcglm} fitted model object.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

