% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_anova.R
\name{anova.mcglm}
\alias{anova.mcglm}
\title{ANOVA method for McGLMs.}
\usage{
\method{anova}{mcglm}(object)
}
\arguments{
\item{object}{an object of class mcglm, usually, a result of a call to \code{mcglm}.}
}
\description{
ANOVA method for McGLMS.
}

