% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_coef.R
\name{coef.mcglm}
\alias{coef.mcglm}
\title{Extract model coefficients for mcglm class}
\usage{
\method{coef}{mcglm}(object, response = c(NA, 1:length(object$beta_names)),
  which = c("beta", "tau", "power", "correlation"))
}
\arguments{
\item{object}{An object of "mcglm" class.}

\item{response}{A numeric or vector specyfing for which response variable the coefficients
should be returned.}

\item{which}{A string specyfing which coefficients should be returned. Options are "all",
"regression", "correlation", "power" and "tau".}
}
\value{
A vector of coefficients.
}
\description{
coefficient is a function which extracts model coefficients from objects of "mcglm" class.
}

