% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_coef.R
\name{coef.mcglm}
\alias{coef.mcglm}
\title{Extract model coefficients for mcglm class}
\usage{
\method{coef}{mcglm}(object, std.error = FALSE, response = c(NA,
  1:length(object$beta_names)), type = c("beta", "tau", "power",
  "correlation"))
}
\arguments{
\item{object}{An object of mcglm class.}

\item{std.error}{Logical. Returns or not the standard errors.}

\item{response}{A numeric or vector specyfing for which response variables the coefficients
should be returned.}

\item{type}{A string or string vector specyfing which coefficients should be returned.
Options are 'beta', 'tau', 'power', 'tau' and 'correlation'.}
}
\value{
A data.frame with estimates, parameters names, response number and parameters type.
}
\description{
coef.mcglm is a function which extracts model coefficients from objects of mcglm class.
}

