% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{coef.mcglm}
\alias{coef.mcglm}
\title{Extract model coefficients for mcglm class}
\usage{
\method{coef}{mcglm}(object, std.error = FALSE, response = c(NA,
  1:length(object$beta_names)), type = c("beta", "tau", "power",
  "correlation"), ...)
}
\arguments{
\item{object}{An object of \code{mcglm} class.}

\item{std.error}{Logical. If \code{TRUE} returns the standard errors
of the estimates. Default is \code{FALSE}.}

\item{response}{A numeric vector specyfing for which response
variables the coefficients should be returned.}

\item{type}{A string vector (can be 1 element length) specyfing which
coefficients should be returned. Options are \code{"beta"},
\code{"tau"}, \code{"power"}, \code{"tau"} and
\code{"correlation"}.}

\item{...}{additional arguments affecting the summary produced. Note
    that there is no extra options for mcglm object class.}
}
\value{
A \code{data.frame} with parameters names, estimates,
    response number and parameters type.
}
\description{
\code{coef.mcglm} is a function which extracts model
    coefficients from objects of \code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

