% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_confint.mcglm.R
\name{confint.mcglm}
\alias{confint.mcglm}
\title{Confidence Intervals for mcglm}
\usage{
\method{confint}{mcglm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted mcglm object.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of
number or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments affecting the confidence interval produced.
Note that there is no extra options for
mcglm object class.}
}
\value{
A data.frame with confidence intervals, parameters names,
response number and parameters type.
}
\description{
Computes confidence intervals for parameters in a fitted mcglm.
}

