% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{confint.mcglm}
\alias{confint.mcglm}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{mcglm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted \code{mcglm} object.}

\item{parm}{specifies for which parameters are to be given
confidence intervals, either a vector of numbers or a vector of
strings. If missing, all parameters are considered.}

\item{level}{the nominal confidence level.}

\item{...}{additional arguments affecting the confidence interval
produced. Note that there is no extra options for \code{mcglm}
object class.}
}
\value{
A \code{data.frame} with confidence intervals, parameters
    names, response variable number and parameters type.
}
\description{
Computes confidence intervals for parameters in a fitted
    \code{mcglm} model.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
