% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ess.R
\name{ESS}
\alias{ESS}
\title{Extract Generalized Error Sum of Squares (ESS)}
\source{
Bonat, W. H. (2016). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, submitted.

Wang, M. (2014). Generalized Estimating Equations in Longitudinal Data
Analysis: A Review and Recent Developments. Advances in Statistics, 1(1)1--13.
}
\usage{
ESS(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{verbose}{logical. Print or not the ESS value.}
}
\value{
Returns the value of generalized error sum of squares(ESS).
}
\description{
Extract the generalized error sum of squares for objects of
\code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\seealso{
\code{plogLik}, \code{pAIC} and \code{pKLIC}.
}

