% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mcglm.R
\name{fit_mcglm}
\alias{fit_mcglm}
\title{Chaser and reciprocal likelihood algorithms.}
\usage{
fit_mcglm(list_initial, list_link, list_variance, list_covariance, list_X,
  list_Z, list_offset, list_Ntrial, list_power_fixed, list_sparse, y_vec,
  correct = TRUE, max_iter, tol = 0.001, method = "rc", tunning = 0,
  verbose)
}
\arguments{
\item{list_initial}{A list of initial values for regression and covariance parameters.}

\item{list_link}{A list of link function names.}

\item{list_variance}{A list of variance function names.}

\item{list_covariance}{A list of covariance function names.}

\item{list_X}{A list of design matrices.}

\item{list_Z}{A list of knowm matrices to used on the matrix linear predictor.}

\item{list_offset}{A list of offset values.}

\item{list_Ntrial}{A list of number of trials, useful only when analysis binomial data.}

\item{list_power_fixed}{A list of logicals indicating if the power parameters should be estimated
or not.}

\item{list_sparse}{A list of logicals indicating if the matrices should be set up as sparse
matrices. This argument is useful only when using exponential-matrix covariance link function.
In the other cases the algorithm detects automatically if the matrix should be sparse or not.}

\item{y_vec}{A vector of the response variables stacked.}

\item{correct}{A logical indicating if the algorithm will use the correction term or not.}

\item{max_iter}{Maximum number of iterations.}

\item{tol}{A numeric spcyfing the tolerance.}

\item{method}{A string specyfing the method used to fit the models (\code{chaser} or \code{rc}).}

\item{tunning}{A numeric value in general near zero for the rc method and near 1 for the chaser
method. This argument control the step-length.}

\item{verbose}{A logical if TRUE print the values of the covariance parameters used on each iteration.}
}
\value{
A list with estimated regression and covariance parameters.
}
\description{
This function implements the two manin algorithms used to fitting McGLMs. The chaser
and the reciprocal likelihood algorithms. In general the chaser algorithm is faster than the
reciprocal likelihood and should be preferred.
}

