% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{fitted.mcglm}
\alias{fitted.mcglm}
\title{Fitted Values}
\usage{
\method{fitted}{mcglm}(object, ...)
}
\arguments{
\item{object}{A fitted object of class \code{mcglm}.}

\item{...}{Additional arguments. Currently ignored and included for
compatibility with the generic
\code{\link[stats]{fitted}} function.}
}
\value{
A numeric matrix of fitted values. Rows correspond to observations
and columns correspond to response variables.
}
\description{
Extract fitted (mean) values from a fitted \code{mcglm} model.
For multivariate responses, fitted values are returned in matrix
form, with one column per response variable.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
