% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{fitted.mcglm}
\alias{fitted.mcglm}
\title{Extract Model Fitted Values of McGLM}
\usage{
\method{fitted}{mcglm}(object, ...)
}
\arguments{
\item{object}{An object of \code{mcglm} class.}

\item{...}{additional arguments affecting the summary produced. Note
that there is no extra options for \code{mcglm} object class.}
}
\value{
Depending on the number of response variable, the function
    \code{fitted.mcglm} returns a vector (univariate models) or a
    matrix (multivariate models) of fitted values.
}
\description{
Extract fitted values for objects of \code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

