% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{fitted.mcglm}
\alias{fitted.mcglm}
\title{Fitted Values}
\usage{
\method{fitted}{mcglm}(object, ...)
}
\arguments{
\item{object}{an object of \code{mcglm} class.}

\item{...}{additional arguments affecting the summary produced. Note
that there is no extra options for \code{mcglm} object class.}
}
\value{
A matrix with fitted values.
}
\description{
Extract fitted values for objects of \code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

