% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_gof.R
\name{gof}
\alias{gof}
\title{Extract Measures of Goodness-of-Fit}
\usage{
gof(object)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}
}
\value{
Returns a data frame contains gof's.
}
\description{
Extract the plogLik, pAIC, pKLIC and ESS for an object
of \code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\seealso{
\code{plogLik}, \code{pAIC}, \code{pKLIC}, \code{ESS}.
}

