% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_gof.R
\name{gof}
\alias{gof}
\title{Measures of Goodness-of-Fit}
\source{
Bonat, W. H. (2016). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, submitted.

Wang, M. (2014). Generalized Estimating Equations in Longitudinal Data
Analysis: A Review and Recent Developments. Advances in Statistics, 1(1)1--13.
}
\usage{
gof(object)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}
}
\value{
Returns a data frame containing goodness-of-fit measures.
}
\description{
Extract the pseudo Gaussian log-likelihood (plogLik),
pseudo Akaike Information Criterion (pAIC), pseudo Kullback-Leibler
Information Criterion (pKLIC) and Error Sum of Squares (ESS)
for objects of \code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\seealso{
\code{plogLik}, \code{pAIC}, \code{pKLIC}, \code{ESS}.
}

