% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_build_sigmab.R
\name{mc_build_sigma_between}
\alias{mc_build_sigma_between}
\alias{mc_derivative_sigma_between}
\title{Build the correlation matrix between response variables}
\usage{
mc_build_sigma_between(rho, n_resp, inverse = FALSE)

mc_derivative_sigma_between(n_resp)
}
\arguments{
\item{rho}{A numeric vector.}

\item{n_resp}{A numeric.}

\item{inverse}{Logical.}
}
\value{
A list with sigmab and its derivatives with respect to rho.
}
\description{
This function builds the correlation matrix between
    response variable, its inverse and derivatives.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
