% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_cross_sensitivity.R
\name{mc_cross_sensitivity}
\alias{mc_cross_sensitivity}
\title{Cross-sensitivity}
\usage{
mc_cross_sensitivity(Product_cov, Product_beta,
  n_beta_effective = length(Product_beta))
}
\arguments{
\item{Product_cov}{A list of matrices.}

\item{Product_beta}{A list of matrices.}

\item{n_beta_effective}{Numeric. Effective number of regression
parameters.}
}
\value{
The cross-sensitivity matrix. Equation (10) of Bonat and
    Jorgensen (2016).
}
\description{
Compute the cross-sensitivity matrix between regression
    and covariance parameters.  Equation 10 of Bonat and Jorgensen
    (2015).
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
