% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_id.R
\name{mc_id}
\alias{mc_id}
\title{Independent Model Structure}
\source{
Bonat, W. H. (2016). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, submitted.
}
\usage{
mc_id(data)
}
\arguments{
\item{data}{the data set to be used.}
}
\value{
A list of matrix.
}
\description{
Builds an identity matrix to be used as a component of
the matrix linear predictor. It is in general the first component of
the matrix linear predictor, a kind of intercept matrix.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\seealso{
\code{mc_dist}, \code{mc_ma}, \code{mc_rw} and \code{mc_mixed}.
}

