% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_influence.R
\name{mc_influence}
\alias{mc_influence}
\title{Influence measures}
\usage{
mc_influence(object, id)
}
\arguments{
\item{object}{An object of mcglm class.}

\item{id}{a vector which identifies the clusters. The length and order of id should be the
same as the number of observations. Data are assumed to be sorted so that observations on a cluster
are contiguous rows for all entities in the formula.}
}
\value{
A matrix. Note that the function assumes that the data are in the correct order.
}
\description{
Compute influence measures for multivariate covariance generalized linear models.
Leverage, DFBETA and Cook's distance for unit sample and observations.
}

