% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ma.R
\name{mc_ma}
\alias{mc_ma}
\title{Moving Average Models Structure}
\source{
Bonat, W. H. (2016). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, submitted.
}
\usage{
mc_ma(id, time, data, order = 1)
}
\arguments{
\item{id}{subject index. Note that this structure was designed to deal with longitudinal data.}

\item{time}{index indicating the time.}

\item{data}{data set.}

\item{order}{order of the random walk model.}
}
\value{
A matrix of dgCMatrix class.
}
\description{
Builds moving average model matrix of order k.
}
\details{
This function was designed mainly to deal with longitudinal
data, but can also be used for times series analysis. In that case,
the \code{id} argument should contain only one index. It pretends a
longitudinal data taken just for one individual or unit sample. This
function is a simple call of the \code{\link[Matrix]{bandSparse}}.
}

