% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_manova.R
\name{mc_manova}
\alias{mc_manova}
\title{MANOVA-type test for mcglm objects.}
\usage{
mc_manova(object, ...)
}
\arguments{
\item{object}{an object of \code{mcglm} class.}

\item{...}{additional arguments affecting the summary produced. Note
that there is no extra options for mcglm object class.}
}
\value{
MANOVA table for mcglm objects.
}
\description{
IT IS AN EXPERIMENTAL FUNCTION BE CAREFUL!
MANOVA-type test for multivariate covariance generalized
linear models.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
