% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_manova_disp.R
\name{mc_manova_disp}
\alias{mc_manova_disp}
\title{MANOVA-type test for mcglm objects.}
\usage{
mc_manova_disp(object, idx, names, ...)
}
\arguments{
\item{object}{an object of \code{mcglm} class.}

\item{idx}{vector of indexes for parameter tests.}

\item{names}{vector of names to appear in the ANOVA table.}

\item{...}{additional arguments affecting the summary produced. Note
that there is no extra options for mcglm object class.}
}
\value{
MANOVA table for mcglm objects.
}
\description{
IT IS AN EXPERIMENTAL FUNCTION BE CAREFUL!
MANOVA-type test for multivariate covariance generalized
linear models. This function is specify for the dispersion components.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
