% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_quasi_score.R
\name{mc_quasi_score}
\alias{mc_quasi_score}
\title{Quasi-score function}
\usage{
mc_quasi_score(D, inv_C, y_vec, mu_vec, W)
}
\arguments{
\item{D}{A matrix. In general the output from
\code{\link[mcglm]{mc_link_function}}.}

\item{inv_C}{A matrix. In general the output from
\code{\link[mcglm]{mc_build_C}}.}

\item{y_vec}{A vector.}

\item{mu_vec}{A vector.}

\item{W}{Matrix of weights.}
}
\value{
The quasi-score vector, the Sensivity and variability
    matrices.
}
\description{
Compute the quasi-score function, its sensitivy and
    variability matrix.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
