% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_remove_na.R
\name{mc_remove_na}
\alias{mc_remove_na}
\title{Remove NA from Matrix Linear Predictor}
\source{
Bonat, W. H. (2018). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, 84(4):1--30.
}
\usage{
mc_remove_na(matrix_pred, cod)
}
\arguments{
\item{matrix_pred}{a list of known matrices.}

\item{cod}{index indicating the columns should be removed.}
}
\value{
A list of matrices.
}
\description{
The function \code{mc_remove_na} removes NA from each
component of the matrix linear predictor. It is in general used after
the function \code{mc_complete_data}.
}
\seealso{
\code{mc_dglm}, \code{mc_ns}, \code{mc_ma} and \code{mc_rw}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
