% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sensitivity.R
\name{mc_sensitivity}
\alias{mc_sensitivity}
\title{Sensitivity matrix}
\usage{
mc_sensitivity(product, W)
}
\arguments{
\item{product}{A list of matrix.}

\item{W}{Matrix of weights.}
}
\value{
A matrix. The sensitivity matrix associated with the Pearson estimating
    function. The returned object is intended for internal use only.
}
\description{
Compute the sensitivity matrix associated with the
    Pearson estimating function.
}
\details{
This function implements the equation 7 of Bonat and
    Jorgensen (2016).
}
\author{
Wagner Hugo Bonat and Eduardo Elias Ribeiro Jr
}
\keyword{internal}
