% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sensitivity.R
\name{mc_sensitivity}
\alias{mc_sensitivity}
\title{Sensitivity matrix}
\usage{
mc_sensitivity(product)
}
\arguments{
\item{product}{A list of matrix.}
}
\value{
The sensitivity matrix associated with the Pearson estimating
    function.
}
\description{
Compute the sensitivity matrix associated with the
    Pearson estimating function.
}
\details{
This function implements the equation 7 of Bonat and
    Jorgensen (2016).
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
