% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_transform_list_bdiag.R
\name{mc_transform_list_bdiag}
\alias{mc_transform_list_bdiag}
\title{Auxiliar function to compute the derivatives of the C matrix.}
\usage{
mc_transform_list_bdiag(list_mat, mat_zero, response_number)
}
\arguments{
\item{list_mat}{A list of matrices.}

\item{mat_zero}{A list of zero matrices. In general the output of
\code{link[mcglm]{mc_build_bdiag}}.}

\item{response_number}{A numeric specifying the response variable
number.}
}
\value{
A list of block-diagonal matrices.
}
\description{
This function take a list of matrices and return a list
    of block-diagonal matrices, where the original matrices are one
    block non-zero of the matrix.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}

