% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_twin.R
\name{mc_twin}
\alias{mc_twin}
\title{Twin Models Structure}
\source{
Bonat, W. H. (2016). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, submitted.
}
\usage{
mc_twin(id, twin.id, type, replicate = NULL, structure, data)
}
\arguments{
\item{id}{name of the column (string) containing the twin index.
It should be the same index (number) for both twins.}

\item{twin.id}{name of the column (string) containing the twin index
inside the pair. In general 1 for the first twin and 2 for the second
twin.}

\item{type}{name of the column (string) containing the indication of
the twin as mz or dz. It should be a factor with only two levels mz and dz.}

\item{replicate}{name of the column (string) containing the index
for more than one observation taken at the same twin pair. It is used
for example in twin longitudinal studies. In that case, the replication
column should contain the time index.}

\item{structure}{model type options are A, C and D. In general, one
model will use more than one component like AC or AD models. The user
should also use the \code{mc_id} function to create the E component
of the twin model. See example for details.}

\item{data}{data set.}
}
\value{
A matrix of \code{dgCMatrix} class.
}
\description{
The function \code{mc_twin} helps to build the components
of the matrix linear predictor associated with ACDE models for
analysis of twin data.
}
\examples{
id <- rep(1:5, each = 4)
id.twin <- rep(1:2, 10)
type <- c(rep("mz",12),rep("dz",8))
replication <- rep(rep(1:2, each = 2),5)
data <- data.frame(id, id.twin, type,replication)
mc_twin(twin.id = "id.twin", id = "id", type = "type",
structure = "A", replicate = "replication", data = data)
data2 <- data[order(data$id.twin),]
mc_twin(id = "id", twin.id = "id.twin", type = "type",
structure = "A", replicate = "replication", data = data2)
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\seealso{
\code{mc_id}, \code{mc_dist}, \code{mc_car},
\code{mc_rw}, \code{mc_ns}, \code{mc_dglm} and \code{mc_mixed}.
}

