% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_updatedBeta.R
\name{mc_updateBeta}
\alias{mc_updateBeta}
\title{Updated regression parameters}
\usage{
mc_updateBeta(list_initial, betas, information, n_resp)
}
\arguments{
\item{list_initial}{A list of initial values.}

\item{betas}{A vector with actual regression parameters values.}

\item{information}{A list with information about the number of
parameters in the model. In general the output from
\link{mc_getInformation}.}

\item{n_resp}{A numeric specyfing the number of response variables.}
}
\value{
A list with updated values of the regression parameters.
}
\description{
This function update a list of regression parameters. It
    will be useful only inside the fitting algorithm.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
