% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_variability.R
\name{mc_variability}
\alias{mc_variability}
\title{Variability matrix}
\usage{
mc_variability(sensitivity, product, inv_C, C, res)
}
\arguments{
\item{sensitivity}{A matrix. In general the output from
\code{mc_sensitivity}.}

\item{product}{A list of matrix.}

\item{inv_C}{A matrix. In general the output from \code{mc_build_C}.}

\item{C}{A matrix. In general the output from \code{mc_build_C}.}

\item{res}{A vector. The residuals vector, i.e. (y_vec - mu_vec).}
}
\value{
The variability matrix associated witht the Pearson
    estimating function.
}
\description{
Compute the variability matrix associated with the
    Pearson estimating function.
}
\details{
This function implements the equation 8 of Bonat and
    Jorgensen (2016).
}
\author{
Wagner Hugo Bonat and Eduardo Elias Ribeiro Jr
}
\keyword{internal}
