% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcglm.R
\name{mcglm}
\alias{mcglm}
\title{Fitting Multivariate Covariance Generalized Linear Models (McGLM)}
\usage{
mcglm(linear_pred, matrix_pred, link, variance, covariance, offset, Ntrial,
  power_fixed, data, control_initial, control_algorithm = list(),
  check_convergence = FALSE)
}
\arguments{
\item{linear_pred}{A list of formula see \code{\link[stats]{formula}} for details.}

\item{matrix_pred}{A list of known matrices to be used on the matrix linear predictor. Details
can be obtained on \code{\link[mcglm]{mc_matrix_linear_predictor}}.}

\item{link}{A list of link functions names, see \code{\link[mcglm]{mc_link_function}} for details.}

\item{variance}{A list of variance functions names, see \code{\link[mcglm]{mc_variance_function}}
for details.}

\item{covariance}{A list of covariance link functions names, current options are: identity, inverse
and exponential-matrix (expm).}

\item{offset}{A list with values of offset values if any.}

\item{Ntrial}{A list with values of the number of trials on Bernoulli experiments. It is useful only
for binomialP and binomialPQ variance functions.}

\item{power_fixed}{A list of logicals indicating if the values of the power parameter should be
estimated or not.}

\item{control_initial}{A list of initial values for the fitting algorithm. See details below.}

\item{control_algorithm}{A list of arguments to be passed for the fitting algorithm. See
\code{\link[mcglm]{fit_mcglm}} for details.}

\item{check_convergence}{Logical if TRUE provide a convergence check.}
}
\value{
mcglm returns an object of class "mcglm".
}
\description{
\code{mcglm} is used to fit multivariate covariance generalized linear models.
The models are specified by a set of lists giving a symbolic description of the linear predictor and
specyfing the link, variance and covariance functions.
}

