% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_KLIC.R
\name{pKLIC}
\alias{pKLIC}
\title{Pseudo Kullback-Leibler Information Criterion}
\source{
Bonat, W. H. (2018). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, 84(4):1--30.
}
\usage{
pKLIC(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{verbose}{logical. Print or not the pKLIC value.}
}
\value{
Returns the value of the pseudo Kullback-Leibler information
criterion.
}
\description{
Extract the pseudo Kullback-Leibler information criterion
(pKLIC) for objects of \code{mcglm} class.
}
\seealso{
\code{gof}, \code{plogLik}, \code{ESS}, \code{pAIC},
\code{GOSHO} and \code{RJC}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
