% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plogLik.R
\name{plogLik}
\alias{plogLik}
\title{Gaussian Pseudo-loglikelihood}
\usage{
plogLik(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mcglm} class.}

\item{verbose}{logical. Print or not the plogLik value.}
}
\value{
Returns the value of the Gaussian pseudo-loglikelihood.
}
\description{
Extract the Gaussian pseudo-loglikelihood (plogLik)
value for objects of \code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

