% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{plot.mcglm}
\alias{plot.mcglm}
\title{Default Multivariate Covariance Generalized Linear models plotting}
\usage{
\method{plot}{mcglm}(x, type = "residuals", ...)
}
\arguments{
\item{x}{a fitted \code{mcglm} object.}

\item{type}{Specify which graphical analysis will be performed.
Options are: \code{"residuals"}, \code{"influence"} and
\code{"algorithm"}.}

\item{...}{additional arguments affecting the plot produced. Note
that there is no extra options for mcglm object class.}
}
\description{
takes a fitted \code{mcglm} object and do plots based on
    residuals, influence diagnostic measures and algorithm check.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

