% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{plot.mcglm}
\alias{plot.mcglm}
\title{Default Multivariate Covariance Generalized Linear Models
    plotting}
\usage{
\method{plot}{mcglm}(x, type = "residuals", ...)
}
\arguments{
\item{x}{a fitted \code{mcglm} object.}

\item{type}{specify which graphical analysis will be performed.
Options are: \code{"residuals"} and \code{"algorithm"}.}

\item{...}{additional arguments affecting the plot produced. Note
that there is no extra options for mcglm object class.}
}
\value{
The function \code{plot.mcglm} was designed to offer a fast
residuals analysis based on the Pearson residuals. Current version
offers a simple Pearson residuals versus fitted values and a quantile
plot. When using \code{algorithm = TRUE} the will plot a summary of
the fitting algorithm show the trajectory or iterations of the fitting
algorithm. The iteractions are shown in terms of values for the model
parameters and also the actually value of the quasi-score and Pearson
estimating functions. Hence, a quickly check of the algorithm
convergence is obtained.
}
\description{
Takes a fitted \code{mcglm} object and do plots based on
    residuals and algorithm check.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

