% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_plot.mcglm.R
\name{plot.mcglm}
\alias{plot.mcglm}
\title{Default Multivariate Covariance Generalized Linear models plotting}
\usage{
\method{plot}{mcglm}(x, type = "residuals", ...)
}
\arguments{
\item{x}{a fitted mcglm object as produced by \code{mcglm()}.}

\item{type}{Specify which graphical analysis will be performed, options are: residuals, influence
and algorithm.}

\item{...}{additional arguments affecting the plot produced. Note that there is no extra options for
mcglm object class.}
}
\description{
takes a fitted mcglm object by \code{mcglm()} and plots, residuals,
influence diagnostic measures and algorithm check.
}

