% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_print.mcglm.R
\name{print.mcglm}
\alias{print}
\alias{print.mcglm}
\title{Print a Multivariate Covariance Generalized Linear Model}
\usage{
print.mcglm(object)
}
\arguments{
\item{object}{fitted model objects of class mcglm as produced by mcglm().}
}
\description{
The default print method for a mcglm object.
}

