% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{print.mcglm}
\alias{print.mcglm}
\title{Print method for Multivariate Covariance Generalized Linear
    Model}
\usage{
\method{print}{mcglm}(x, ...)
}
\arguments{
\item{x}{fitted model objects of class mcglm as produced by mcglm().}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The default print method for a \code{mcglm} object.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

