% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{print.mcglm}
\alias{print.mcglm}
\title{Print}
\usage{
\method{print}{mcglm}(x, ...)
}
\arguments{
\item{x}{fitted model objects of class \code{mcglm} as produced by \code{mcglm()}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The default print method for an object of \code{mcglm} class.
}
\seealso{
\code{summary}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
