% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_residuals.mcglm.R
\name{residuals.mcglm}
\alias{residuals.mcglm}
\title{Residuals for Multivariate Covariance Generalized Linear Models (McGLM)}
\usage{
\method{residuals}{mcglm}(object, type = "raw")
}
\arguments{
\item{object}{An of class mcglm, typically the result of a call to \code{mcglm}.}

\item{type}{the type of residuals which should be returned. The alternatives are: "raw"
(default), "pearson" and "standardized".}
}
\value{
Depending on the number of response variable the function \code{residuals.mcglm} returns
a vector (univariate models) or a matrix (multivariate models) of residuals values.
}
\description{
Compute residuals based on fitting mcglm models.
}

