% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{residuals.mcglm}
\alias{residuals.mcglm}
\title{Residuals}
\usage{
\method{residuals}{mcglm}(object, type = "raw", ...)
}
\arguments{
\item{object}{an object of \code{mcglm} class.}

\item{type}{the type of residuals which should be returned.
Options are: \code{"raw"} (default), \code{"pearson"} and
\code{"standardized"}.}

\item{...}{additional arguments affecting the residuals
produced. Note that there is no extra options for mcglm object
class.}
}
\value{
The function \code{residuals.mcglm} returns a matrix of
residuals values.
}
\description{
Compute residuals for an object of \code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\seealso{
\code{fitted}.
}

