% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_RJC.R
\name{RJC}
\alias{RJC}
\title{Rotnitzky-Jewell Information Criterion}
\source{
Wang, M. (2014). Generalized Estimating Equations in Longitudinal Data
Analysis: A Review and Recent Developments. Advances in Statistics, 1(1)1--13.
}
\usage{
RJC(object, id, verbose = TRUE)
}
\arguments{
\item{object}{an object of \code{mcglm} class.}

\item{id}{a vector which identifies the clusters. The length and
order of \code{id} should be the same as the number of
observations. Data are assumed to be sorted so that observations
on a cluster are contiguous rows for all entities in the formula.}

\item{verbose}{logical. Print or not the RJC value.}
}
\value{
The value of the Rotnitzky-Jewell information criterion.
Note that the function assumes that the data are in the correct order.
}
\description{
Compute the Rotnitzky-Jewell information criterion
for an object of \code{mcglm} class.
WARNINGS: This function is limited to models with ONE response variable.
}
\seealso{
\code{gof}, \code{plogLik}, \code{pAIC}, \code{pKLIC},
\code{ESS} and \code{GOSHO}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
