% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{summary.mcglm}
\alias{summary.mcglm}
\title{Summarizing Multivariate Covariance Generalized Linear Models
    fits.}
\usage{
\method{summary}{mcglm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{mcglm}, usually, a result of a
call to \code{mcglm}.}

\item{...}{additional arguments affecting the summary produced. Note
    the there is no extra options for mcglm object class.}
}
\value{
Print an \code{mcglm} object.
}
\description{
Summary for McGLMs objects.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

