% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_summary.mcglm.R
\name{summary.mcglm}
\alias{summary.mcglm}
\title{Summarizing Multivariate Covariance Generalized Linear Models fits.}
\usage{
summary.mcglm(object)
}
\arguments{
\item{object}{an object of class mcglm, usually, a result of a call to \code{mcglm}.}
}
\description{
Summary for mcglm objects.
}

