% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_vcov.R
\name{vcov.mcglm}
\alias{vcov.mcglm}
\title{Calculate Variance-Covariance matrix for a fitted McGLM object.}
\usage{
\method{vcov}{mcglm}(object)
}
\arguments{
\item{object}{a fitted model mcglm object.}
}
\value{
A variance-covariance matrix.
}
\description{
Returns the variance-covariance matrix for all parameters of a mcglm fitted model object.
}

