% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_S3_methods.R
\name{vcov.mcglm}
\alias{vcov.mcglm}
\title{Variance-Covariance Matrix}
\usage{
\method{vcov}{mcglm}(object, ...)
}
\arguments{
\item{object}{an object of \code{mcglm} class.}

\item{...}{additional arguments affecting the summary produced. Note
that there is no extra options for mcglm object class.}
}
\value{
A variance-covariance matrix.
}
\description{
Returns the variance-covariance matrix for an object
of \code{mcglm} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}

